fix: Check if binary is outside the user folder for `Move exe`
By checking if the binary is under the user folder (ex. `/home/user`),
we can make a pretty good guess on whether the binary has the
permissions to move itself to the library. If the binary is outside the
user folder, we disable the `Move exe` feature by default, and add a
warning (but the user can still decide to enable this feature).

This is mainly for packaging efforts like the AUR, where Blender
Launcher is usually installed somewhere like `/usr/bin` or `/bin`, and
thus cannot move itself anyway.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>