fix: Fix issues with `xdg-open` when built with Pyinstaller
Due to a bug/feature in Pyinstaller, `xdg-open` won't work properly by
default. This is due to Pyinstaller changing environment variables at
runtime, breaking what `xdg-open` expects.

As a workaround, before running `xdg-open`, we create a copy of the
environment variables, and remove any variable (except `PATH`) that
contains `tmp` (as noted in
https://github.com/pyinstaller/pyinstaller/issues/3668#issuecomment-742547785).
We then pass that modified copy to `subprocess.run` (thus leaving the
original environment variables unmodified for the rest of Blender
Launcher).

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>